requirements.txt
django
djangorestframework
django-autoslug
django-ckeditor
django-environ
psycopg2
drf-yasg
djangorestframework-simplejwt